Move translation_helper to utils/ dir.
