+++ gslib/command.py
@@ -74,8 +74,6 @@ from gslib.thread_message import FinalMessage
 from gslib.thread_message import MetadataMessage
 from gslib.thread_message import PerformanceSummaryMessage
 from gslib.thread_message import ProducerThreadMessage
-from gslib.translation_helper import AclTranslation
-from gslib.translation_helper import PRIVATE_DEFAULT_OBJ_ACL
 from gslib.ui_controller import MainThreadUIQueue
 from gslib.ui_controller import UIController
 from gslib.ui_controller import UIThread
@@ -94,8 +92,9 @@ from gslib.utils.parallelism_framework_util import ZERO_TASKS_TO_DO_ARGUMENT
 from gslib.utils.rsync_util import RsyncDiffToApply
 from gslib.utils.system_util import IS_WINDOWS
 from gslib.utils.system_util import GetTermLines
+from gslib.utils.translation_helper import AclTranslation
+from gslib.utils.translation_helper import PRIVATE_DEFAULT_OBJ_ACL
 from gslib.wildcard_iterator import CreateWildcardIterator
-
 
 OFFER_GSUTIL_M_SUGGESTION_THRESHOLD = 5
