+++ gslib/commands/cors.py
@@ -28,10 +28,9 @@ from gslib.help_provider import CreateHelpText
 from gslib.storage_url import StorageUrlFromString
 from gslib.storage_url import UrlsAreForSingleProvider
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
-from gslib.translation_helper import CorsTranslation
-from gslib.translation_helper import REMOVE_CORS_CONFIG
 from gslib.utils.constants import NO_MAX
-
+from gslib.utils.translation_helper import CorsTranslation
+from gslib.utils.translation_helper import REMOVE_CORS_CONFIG
 
 _GET_SYNOPSIS = """
   gsutil cors get url
