+++ gslib/commands/label.py
@@ -30,10 +30,10 @@ from gslib.exception import CommandException
 from gslib.exception import NO_URLS_MATCHED_TARGET
 from gslib.help_provider import CreateHelpText
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
-from gslib.translation_helper import LabelTranslation
 from gslib.utils.constants import NO_MAX
 from gslib.utils.constants import UTF8
 from gslib.utils.retry_util import Retry
+from gslib.utils.translation_helper import LabelTranslation
 
 _SET_SYNOPSIS = """
   gsutil label set label-json-file url...
