+++ gslib/commands/lifecycle.py
@@ -27,9 +27,8 @@ from gslib.exception import NO_URLS_MATCHED_TARGET
 from gslib.help_provider import CreateHelpText
 from gslib.storage_url import UrlsAreForSingleProvider
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
-from gslib.translation_helper import LifecycleTranslation
 from gslib.utils.constants import NO_MAX
-
+from gslib.utils.translation_helper import LifecycleTranslation
 
 _GET_SYNOPSIS = """
   gsutil lifecycle get url
