+++ gslib/copy_helper.py
@@ -95,23 +95,14 @@ from gslib.tracker_file import ReadOrCreateDownloadTrackerFile
 from gslib.tracker_file import SERIALIZATION_UPLOAD_TRACKER_ENTRY
 from gslib.tracker_file import TrackerFileType
 from gslib.tracker_file import WriteDownloadComponentTrackerFile
-from gslib.translation_helper import AddS3MarkerAclToObjectMetadata
-from gslib.translation_helper import CopyObjectMetadata
-from gslib.translation_helper import DEFAULT_CONTENT_TYPE
-from gslib.translation_helper import ObjectMetadataFromHeaders
-from gslib.translation_helper import PreconditionsFromHeaders
-from gslib.translation_helper import S3MarkerAclFromObjectMetadata
 from gslib.utils import parallelism_framework_util
-from gslib.utils.cloud_api_helper import GetCloudApiInstance
-from gslib.utils.cloud_api_helper import GetDownloadSerializationData
-from gslib.utils.parallelism_framework_util import AtomicDict
-from gslib.utils.parallelism_framework_util import CheckMultiprocessingAvailableAndInit
-from gslib.utils.parallelism_framework_util import PutToQueueWithTimeout
 from gslib.utils.boto_util import GetJsonResumableChunkSize
 from gslib.utils.boto_util import GetMaxRetryDelay
 from gslib.utils.boto_util import GetNumRetries
 from gslib.utils.boto_util import ResumableThreshold
 from gslib.utils.boto_util import UsingCrcmodExtension
+from gslib.utils.cloud_api_helper import GetCloudApiInstance
+from gslib.utils.cloud_api_helper import GetDownloadSerializationData
 from gslib.utils.constants import DEFAULT_FILE_BUFFER_SIZE
 from gslib.utils.constants import MIN_SIZE_COMPUTE_LOGGING
 from gslib.utils.constants import UTF8
@@ -129,6 +120,9 @@ from gslib.utils.hashing_helper import GetDownloadHashAlgs
 from gslib.utils.hashing_helper import GetUploadHashAlgs
 from gslib.utils.hashing_helper import HashingFileUploadWrapper
 from gslib.utils.metadata_util import ObjectIsGzipEncoded
+from gslib.utils.parallelism_framework_util import AtomicDict
+from gslib.utils.parallelism_framework_util import CheckMultiprocessingAvailableAndInit
+from gslib.utils.parallelism_framework_util import PutToQueueWithTimeout
 from gslib.utils.posix_util import ATIME_ATTR
 from gslib.utils.posix_util import ConvertDatetimeToPOSIX
 from gslib.utils.posix_util import GID_ATTR
@@ -140,6 +134,12 @@ from gslib.utils.system_util import CheckFreeSpace
 from gslib.utils.system_util import GetFileSize
 from gslib.utils.system_util import GetStreamFromFileUrl
 from gslib.utils.system_util import IS_WINDOWS
+from gslib.utils.translation_helper import AddS3MarkerAclToObjectMetadata
+from gslib.utils.translation_helper import CopyObjectMetadata
+from gslib.utils.translation_helper import DEFAULT_CONTENT_TYPE
+from gslib.utils.translation_helper import ObjectMetadataFromHeaders
+from gslib.utils.translation_helper import PreconditionsFromHeaders
+from gslib.utils.translation_helper import S3MarkerAclFromObjectMetadata
 from gslib.utils.unit_util import DivideAndCeil
 from gslib.utils.unit_util import HumanReadableToBytes
 from gslib.utils.unit_util import MakeHumanReadable
