+++ gslib/tests/test_lifecycle.py
@@ -25,8 +25,8 @@ import gslib.tests.testcase as testcase
 from gslib.tests.testcase.integration_testcase import SkipForS3
 from gslib.tests.util import ObjectToURI as suri
 from gslib.tests.util import unittest
-from gslib.translation_helper import LifecycleTranslation
 from gslib.utils.retry_util import Retry
+from gslib.utils.translation_helper import LifecycleTranslation
 
 
 @SkipForS3('Lifecycle command is only supported for gs:// URLs')
