+++ gslib/utils/translation_helper.py
